requirements.txt

Flask==0.10.1
Flask-SQLAlchemy==1.0
Flask-WTF==0.9.3

Flask-Login==0.4.0
backports.pbkdf2==0.1

